Restored the coupled_0.001_impulse example.
Made changes to the input_h.py which is the test
file that is used to test the convection model.

Saved changes to the tutorial notebook.

Created pyrk/conductivity_model.py
to define the 'ConductivityModel' class which
linearly interpolates the thermal conductivity
using celcius.

Applied conductivity_model to flibe.py